Trait openexr::frame_buffer::PixelData

pub unsafe trait PixelData {
    fn pixel_type() -> PixelType;
}

Types that can be inserted into a FrameBuffer as a channel.

Implementing this trait on a type allows the type to be used directly by the library to write data out to and read data in from EXR files.

NOTE: this should only be implemented for types that are bitwise- and semantically-identical to one of the PixelType variants. It has already been implemented for the applicable built-in Rust types, as well as f16 from the Half crate.

Required Methods

fn pixel_type() -> PixelType

Returns which PixelType the implementing type corresponds to.