Struct openexr::FrameBuffer

pub struct FrameBuffer<'a> { /* fields omitted */ }

Points to and describes in-memory image data for reading and writing.

FrameBuffer does not store any image data itself, but instead points to and describes image data elsewhere in memory. Those descriptions are then used by the input and output file types to know where in memory to read from and write to when writing and reading files.

Methods

impl<'a> FrameBuffer<'a>

fn new(width: usize, height: usize) -> Self

Creates an empty frame buffer with the given dimensions in pixels.

FrameBuffer does not store any data, therefore its size in memory is independent of the dimensions specified here.

fn dimensions(&self) -> (usize, usize)

Return the dimensions of the frame buffer.

fn insert_channel<T: PixelData>(&mut self,
name: &str,
fill: f64,
data: &'a [T])
-> &mut Self

Insert a single channel.

The channel will be given the name name and when reading from a file will be filled in with the value fill if there isn't any pixel data for that channel in the file.

data is the memory for the channel and should contain precisely width * height elements, where width and height are the dimensions of the FrameBuffer.

fn insert_channels<T: PixelStruct>(&mut self,
channels: &[(&str, f64)],
data: &'a [T])
-> &mut Self

Insert multiple channels from a slice of structs or tuples.

The number of channels to be inserted is determined by the implementation of the PixelStruct trait on T. channels should contain that number of elements, and each element is a tuple of the channel's name and default fill value.

data is the memory for the channel and should contain precisely width * height elements, where width and height are the dimensions of the FrameBuffer.

unsafe fn insert_raw(&mut self,
name: &str,
type_: PixelType,
base: *const c_char,
stride: (usize, usize),
sampling: (c_int, c_int),
fill_value: f64,
tile_coords: (bool, bool))
-> &mut Self

The raw method for inserting a new channel.

This is very unsafe: the other methods should be preferred unless you have a special use-case.

This method corresponds directly to constructing and then inserting a "Slice" in the C++ OpenEXR library. Please see its documentation for details.

fn validate_header_for_output(&self, header: &Header) -> Result<()>