Struct openexr::Header

pub struct Header { /* fields omitted */ }

Represents an OpenEXR file header.

The file header describes the properties of the image, such as image resolution, the channels it contains, custom attributes, etc. It is used both for fetching information about a loaded EXR file and for defining the properties of a file to be written.

Examples

Creating a header for a file that will be written:

use openexr::{Header, PixelType};

Header::new()
    .set_resolution(1920, 1080)
    .add_channel("R", PixelType::FLOAT)
    .add_channel("G", PixelType::FLOAT)
    .add_channel("B", PixelType::FLOAT);

Methods

impl Header

fn new() -> Self

Creates a new header.

fn set_resolution(&mut self, width: u32, height: u32) -> &mut Self

Sets the resolution.

This is really just a shortcut for setting both the display window and data window to (0, 0), (width-1, height-1).

fn set_display_window(&mut self, window: Box2i) -> &mut Self

Sets the display window.

For simple use-cases, it's better to use set_resolution() instead.

fn set_data_window(&mut self, window: Box2i) -> &mut Self

Sets the data window.

For simple use-cases, it's better to use set_resolution() instead.

fn set_pixel_aspect_ratio(&mut self, aspect_ratio: f32) -> &mut Self

Sets the pixel aspect ratio.

fn set_screen_window_center(&mut self, center: (f32, f32)) -> &mut Self

Sets the screen window center.

fn set_screen_window_width(&mut self, width: f32) -> &mut Self

Sets the screen window width.

fn set_line_order(&mut self, line_order: LineOrder) -> &mut Self

Sets the line order.

fn set_compression(&mut self, compression: Compression) -> &mut Self

Sets the compression mode.

fn add_channel(&mut self, name: &str, pixel_type: PixelType) -> &mut Self

Adds a channel.

This is a simplified version of add_channel_detailed(), using some reasonable defaults for the details. Specifically: sampling is set to (1, 1) and p_linear is set to true.

fn add_channel_detailed(&mut self, name: &str, channel: Channel) -> &mut Self

Adds a channel, specifying full details.

fn data_dimensions(&self) -> (u32, u32)

Convenience method for the dimensions of the data window.

fn data_window(&self) -> &Box2i

Access to the data window.

fn display_window(&self) -> &Box2i

Access to the display window.

fn channels<'a>(&'a self) -> ChannelIter<'a>

Returns an iterator over the channels in the header.

fn get_channel<'a>(&'a self, name: &str) -> Option<&'a Channel>

Access channels by name.