Struct openexr::frame_buffer::FrameBuffer

pub struct FrameBuffer<'a> { /* fields omitted */ }

Points to and describes in-memory image data for reading.

Methods

impl<'a> FrameBuffer<'a>

fn new(width: u32, height: u32) -> Self

Creates an empty frame buffer with the given dimensions in pixels.

fn dimensions(&self) -> (u32, u32)

Return the dimensions of the frame buffer.

fn insert_channel<T: PixelData>(
&mut self,
name: &str,
data: &'a [T]
) -> &mut Self

Insert a single channel into the FrameBuffer.

The channel will be given the name name.

data is the memory for the channel and should contain precisely width * height elements, where width and height are the dimensions of the FrameBuffer.

fn insert_channels<T: PixelStruct>(
&mut self,
names: &[&str],
data: &'a [T]
) -> &mut Self

Insert multiple channels from a slice of structs or tuples.

The number of channels to be inserted is determined by the implementation of the PixelStruct trait on T. names should contain the names of each of those channels.

data is the memory for the channel and should contain precisely width * height elements, where width and height are the dimensions of the FrameBuffer.

unsafe fn insert_raw(
&mut self,
name: &str,
type_: PixelType,
base: *const c_char,
stride: (usize, usize),
sampling: (c_int, c_int),
fill_value: f64,
tile_coords: (bool, bool)
) -> &mut Self

The raw method for inserting a new channel.

This is very unsafe: the other methods should be preferred unless you have a special use-case.

This method corresponds directly to constructing and then inserting a "Slice" in the C++ OpenEXR library. Please see its documentation for details.